pyproject.toml
[project]
name = "chaimcp"
version = "0.1.0"
description = "ChaiMCP Server - Connect your Agentic IDE to Chia Blockchain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.scripts]
chaimcp = "chaimcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chaimcp"]
